Renamed instances of "MftReader" to "MFTSerializer"
