+++ src/MFTSerializer/Constants.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MFTSerializer
+{
+    public class Constants
+    {
+        public const string APP_NAME = "MFTSerializer";
+        public const string APP_CHANNEL = "Release";
+        public const string APP_VERSION = "1.0";
+        public const string APP_SIGNATURE = $"{APP_NAME} - {APP_CHANNEL} {APP_VERSION}";
+        public static string APP_URL = "https://github.com/Scicrop/MftReader";
+    }
+}
