+++ src/MFTSerializer/PInvokeWin32.cs
