+++ src/MFTSerializer/Properties/AssemblyInfo.cs
