+++ src/MFTSerializer/appsettings.json
