+++ src/MFTSerializer/scicrop-MFTSerializer-icon.ico
