--- src/MftReader/Constants.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace MftReader
-{
-    public class Constants
-    {
-        public const string APP_NAME = "MftReader";
-        public const string APP_VERSION = "1.2.0";
-        public const string APP_SIGNATURE = APP_NAME+" "+APP_VERSION;
-        public static string APP_URL = "https://github.com/Scicrop/MftReader";
-    }
-}
